Fix a race condition between a file truncate (change of file size) and pending writes extending the file.
The issue was coming from the fact that the wr_block was not blocked by the setattr. As a consequence
any response associated with a write that was extending the file might overwrite the size sets by the
ftruncate that was triggered after a write (remember that rozofs was doing write behind).